requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-extensions==3.2.3
djangorestframework==3.15.2
filetype==1.2.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.44
langchain-google-genai==2.1.0
langgraph==0.3.8
langgraph-checkpoint==2.0.19
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.56
langsmith==0.3.13
msgpack==1.1.0
orjson==3.10.15
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zstandard==0.23.0
